Re-organize code
