+++ LinkedIdeas/CanvasView.swift
@@ -8,21 +8,6 @@
 
 import Cocoa
 
-protocol LinkedIdeasDocument {
-  var concepts: [Concept] { get }
-  var links: [Link] { get }
-  
-  var observer: DocumentObserver? { get set }
-  
-  func saveConcept(concept: Concept)
-  func removeConcept(concept: Concept)
-  
-  func saveLink(link: Link)
-  func removeLink(link: Link)
-  
-  func changeConceptPoint(concept: Concept, fromPoint: NSPoint, toPoint: NSPoint)
-}
-
 class CanvasView: NSView, Canvas, DocumentObserver {
   var document: LinkedIdeasDocument! {
     didSet { document.observer = self }
