+++ LinkedIdeas/DocumentData.swift
@@ -1,0 +1,30 @@
+//
+//  DocumentData.swift
+//  LinkedIdeas
+//
+//  Created by Felipe Espinoza on 01/05/2016.
+//  Copyright © 2016 Felipe Espinoza Dev. All rights reserved.
+//
+
+import Foundation
+
+class DocumentData: NSObject, NSCoding {
+  var readConcepts: [Concept]?
+  var readLinks: [Link]?
+  var writeConcepts: [Concept]?
+  var writeLinks: [Link]?
+  
+  override init() {
+    super.init()
+  }
+  
+  required init?(coder aDecoder: NSCoder) {
+    readConcepts = aDecoder.decodeObjectForKey("concepts") as! [Concept]?
+    readLinks = aDecoder.decodeObjectForKey("links") as! [Link]?
+  }
+  
+  func encodeWithCoder(aCoder: NSCoder) {
+    aCoder.encodeObject(writeConcepts, forKey: "concepts")
+    aCoder.encodeObject(writeLinks, forKey: "links")
+  }
+}
