+++ LinkedIdeas/Mode.swift
@@ -1,0 +1,15 @@
+//
+//  Mode.swift
+//  LinkedIdeas
+//
+//  Created by Felipe Espinoza on 01/05/2016.
+//  Copyright © 2016 Felipe Espinoza Dev. All rights reserved.
+//
+
+import Foundation
+
+enum Mode: String {
+  case Select = "select"
+  case Concepts = "concepts"
+  case Links = "links"
+}
